refactor: move login/logout/register functions from useUser hook to useAuth
