+++ src/hooks/useUser.ts
@@ -1,157 +1,27 @@
-import { useNavigate } from 'react-router-dom'
 import {
   type Address,
-  type MyCustomerDraft,
-  type MyCustomerSignin,
   type MyCustomerUpdateAction,
   type AddressDraft,
 } from '@commercetools/platform-sdk'
 import { SYSTEM_MESSAGES } from '@/utils/constants'
-import {
-  type UserRegisterPayloadType,
-  type UserUpdatePayloadType,
-} from '@/Models/Models'
+import { type UserUpdatePayloadType } from '@/Models/Models'
 import isErrorResponse from '@/utils/isErrorResponse'
 import useAuth from './useAuth'
 
 export default function useUser(
-  params: Pick<
-    ReturnType<typeof useAuth>,
-    | 'authApiRoot'
-    | 'currentUserRef'
-    | 'setCurrentUser'
-    | 'login'
-    | 'logout'
-    | 'register'
-    | 'isAuthenticated'
-  > & {
+  params: ReturnType<typeof useAuth> & {
     systemMessage: (msg: string, isSuccess: boolean) => void
     setIsFetching: (isFetching: boolean) => void
   },
 ) {
-  const navigate = useNavigate()
   const {
     authApiRoot,
     currentUserRef,
     setCurrentUser,
-    login,
-    logout,
-    register,
     isAuthenticated,
     systemMessage,
     setIsFetching,
   } = params
-
-  const userLogin = async (data: MyCustomerSignin) => {
-    try {
-      setIsFetching(true)
-      const { customer } = await login(data)
-      if (customer) {
-        systemMessage(
-          `${SYSTEM_MESSAGES.LOGIN_SCSS} ${customer.firstName}`,
-          true,
-        )
-        navigate('/')
-      }
-    } catch (e) {
-      systemMessage(SYSTEM_MESSAGES.LOGIN_FAIL, false)
-      throw e
-    } finally {
-      setIsFetching(false)
-    }
-  }
-
-  const userLogout = () => {
-    if (!isAuthenticated) {
-      return
-    }
-    logout()
-    systemMessage(`${SYSTEM_MESSAGES.LOGOUT_SCSS}`, true)
-  }
-
-  const deriveCustomerPayload = (data: UserRegisterPayloadType) => {
-    const {
-      email,
-      password,
-      firstName,
-      lastName,
-      street: shippingAddrStreetName,
-      bldng: shippingAddrStreetNumber,
-      zipCode: shippingAddrPostalCode,
-      city: shippingAddrCity,
-      country: shippingAddrCountry,
-      billingStreet: billAddrStreetName,
-      billingBldng: billAddrStreetNumber,
-      billingZipCode: billAddrPostalCode,
-      billingCity: billAddrCity,
-      billingCountry: billAddrCountry,
-      dateOfBirth,
-      isBillingAddressSame,
-      setDefaultShipAddress: setDefaultShippingAddress,
-      setDefaultBillingAddress,
-    } = data
-
-    const shippingAddress = {
-      streetName: shippingAddrStreetName,
-      streetNumber: shippingAddrStreetNumber,
-      postalCode: shippingAddrPostalCode,
-      city: shippingAddrCity,
-      country: shippingAddrCountry,
-    }
-    const billingAddress = {
-      streetName: billAddrStreetName,
-      streetNumber: billAddrStreetNumber,
-      postalCode: billAddrPostalCode,
-      city: billAddrCity,
-      country: billAddrCountry,
-    }
-
-    const addresses = [shippingAddress]
-
-    const indexOf = (t: (typeof addresses)[number]) => {
-      const index = addresses.indexOf(t)
-      return index !== -1 ? index : undefined
-    }
-
-    let defaultShippingAddress
-    let defaultBillingAddress
-
-    if (setDefaultShippingAddress) {
-      defaultShippingAddress = indexOf(shippingAddress)
-    }
-
-    if (isBillingAddressSame) {
-      defaultBillingAddress = indexOf(shippingAddress)
-    } else {
-      addresses.push(billingAddress)
-      if (setDefaultBillingAddress) {
-        defaultBillingAddress = indexOf(billingAddress)
-      }
-    }
-
-    return {
-      email,
-      password,
-      firstName,
-      lastName,
-      dateOfBirth,
-      addresses,
-      defaultShippingAddress,
-      defaultBillingAddress,
-    } satisfies MyCustomerDraft
-  }
-
-  const userRegister = async (data: UserRegisterPayloadType) => {
-    try {
-      setIsFetching(true)
-      await register(deriveCustomerPayload(data))
-      systemMessage(SYSTEM_MESSAGES.REGISTER_SCSS, false)
-    } catch (e) {
-      systemMessage(SYSTEM_MESSAGES.REGISTER_FAIL, true)
-    } finally {
-      setIsFetching(false)
-    }
-  }
 
   const myCustomerUpdateActions = async (
     ...actions: MyCustomerUpdateAction[]
@@ -331,9 +201,6 @@ export default function useUser(
   }
 
   return {
-    userLogin,
-    userLogout,
-    userRegister,
     setDefaultAddress,
     setAddress,
     addAddress,
